ansible: run application as "nginx" UID
Prior to this commit, the chacra application would run as the "admin"
UID ( {{ ansible_ssh_user }} ). This account has sudo privileges and it
will be safer to run the application under a UID that cannot sudo.

I've selected the "nginx" UID because it is one that we create
explicitly, so it's guaranteed to be present on both CentOS and Ubuntu.