ansible: remove celery role and move it into common
We must install rabbitmq-server after setting the hostname. If we do not
do this then rabbitmq-server will always try to connect using the
hostname of the machine it had at provision time. This is a problem when
restarting rabbitmq-server as it'll fail to start because that old
hostname is no longer resolveable.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>