async/rpm: Create repodata repo root dir for openSUSE/SLE
On openSUSE and SLE, the repository directory structure is different
to the currently used structure. On openSUSE, we expect something
like:

noarch/
repodata/
src/
x86_64/

so the repodata/ dir is on the same level than the other dirs.
This is currently not the case so fix this if the distro is set to
"opensuse".

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>