asynch: use 'pyamqp' broker to avoid librabbitmq 2.0
Without this fix, chacra-celery fails when scheduling its turnon
ping callback to shaman to announce its healthy presence (and presumably
other callbacks would fail if allowed) with
https://github.com/celery/celery/issues/4693.

It's not at all clear what the developers intend the true answer for
https://github.com/celery/celery/issues/4693 to be; there seems to be
some sentiment that librabbitmq is no longer advantageous for speed,
and that a pure Python amqp library is OK.  Apparently it was always
the fallback if librabbitmq (the Python binding for rabbitmq-c) was
not present and 'amqp://' was specified as the broker protocol.  Using
'pyamqp://' forces the pure Python implementation to be used, and the error
goes away.  Is this right?  It'd be great to know.

FWIW, https://github.com/celery/librabbitmq/issues/137 also seems to hint
that librabbitmq is not the good path.

Signed-off-by: Dan Mick <dmick@redhat.com>