debian.py: Make sure subprocess is complete before checking returncode
This was always wrong; there was always a race between the subprocess
and the returncode check.  Something about moving to Py3 has
exposed this (perhaps the implementation had a mistaken wait, or
perhaps the new implementation is just faster.)

Anyway, since we're already using communicate(), move it to before the
returncode check, as it will wait for process completion.

Signed-off-by: Dan Mick <dmick@redhat.com>